Video playback with Studio h264 bitstream parsing (#1)
For interim review. If accepted, the [jhurliman/video
PR](https://github.com/foxglove/studio/pull/5806) will be updated.

This approach uses a message structure identical to
foxglove.CompressedImage. It prompts the question of whether a new
message definition is needed, since either foxglove.CompressedVideo or
foxglove.CompressedImage can be used with no change to the Studio code.

Only H264 video is currently supported, using a format string of "h264".
More codec support can be added in a straightforward way.

The bitstream parsing code does not do any allocations or copies, and
does a very lightweight pass per-frame. The first-time initialization
when the first keyframe is encountered is slightly more involved but
should still be very fast.