Re-run topic alias functions when global variables change (#7074)
This change updates the `TopicAliasingPlayer` to reprocess the last
player state when global variables lead to new topic aliases. Prior to
this change, updating the global variables while paused would not
produce any feedback to the UI that the alias functions had run. It
would only update the state when new state was emitted from the wrapped
player (i.e. during playback or preloading).

This change overhauls the internal architecture of
`TopicAliasingPlayer`. Prior to this change, there was a single state
processing entry point `aliasPlayerState` which was responsible for
creating a set of aliases and then applying those to the player state. I
have changed this to separate the creating of aliases and the applying
into two distinct workflows each with their own memoization. The
processing of alias functions (by providing topics and global variables
as inputs) is now the responsibility of the `StateProcessorFactory`.
Instances of this class memoize the previously processed aliases and
avoid making new processors if the set of aliases is unchanged. This is
important because `setGlobalVariables` can get called often while topic
alias functions may not use any of the updated variables. In these
instances we do not want to be invaliding our state processing or
re-computing state modifications.

The second class is `StateProcessor`. This instance is responsible for
applying a set of aliases to player state and subscriptions. It manages
memoization on the various fields of player state to avoid re-processing
those if they are unchanged in the player state.

With these two classes, the `TopicAliasingPlayer` can opportunistically
evaluate changes to alias inputs and determine if they produced a new
`StateProcessor`. If there is a new state processor (because aliases
changed), then we can re-process the previous player state.

Memoization which used to be global is now isolated to instances of the
two classes specific for their workflows.