Fix incorrect removal of unused topics from preload cache (#7128)
**User-Facing Changes**
Fix removal of unused topics of the preloaded topics cache 

**Description**
When the block loader cache exceeds its limit, we attempt to remove
already preloaded topics that are not needed anymore from existing
blocks. We were however only checking for topic name equality and did
not take into account if the subscription payload (e.g. message slicing
fields) are different. This PR fixes it.