Add setting to hide/show state transition points (#7061)
When working with high-frequency data, the number of state datapoints
can overwhelm our current chart architecture. This change updates the
the dataset creation routines to drop datapoints if they are unchanged
from the last state value. This greatly reduces the number of datapoints
for typical workloads where the state does not transition every single
message in a high-frequency dataset.

When the "show points" setting is false (the new default), we are able
to run the above optimization to cull un-changed points and reduce the
output before we sent it to the chat worker for rendering. This avoids
slow render calls for all of the dots.