Add defer function to foxglove/den/disposable (#7192)
Add a `defer` function which behaves like the `defer` function from
golang and defers the execution of the passed function until the scope
exists. This makes use of the new `Disposables` spec. See comment in
code for more details and usage.

This function is useful if you want to ensure some logic runs regardless
of how the execution of the scope ends. Previously you would accomplish
this via a `try/finally` which would add indentation levels and move the
logic you wanted to defer away from the place where you initialized the
resource you want to defer.