Attempt to resolve `package://` asset URLs relative to URDF url (#7114)
**User-Facing Changes**
attempt to resolve `package://` asset URLs relative to URDF url

**Description**
Often, a URDF contains `package://` references to e.g. mesh files or
xacro includes (converted to package:// URLs by
[xacro-parser](https://www.npmjs.com/package/xacro-parser)). package://
resources can be loaded in 2 different ways:
- Through the player's fetchAsset method (if defined) - e.g. used for
foxglove_websocket
- Through the Desktop app (if ROS_PACKAGE_PATH is defined)

This PR adds a third method in case the first two are not applicable:

```
As last resort to load the package://<pkgName>/<pkgPath> URL, we resolve the package URL to
be relative of the base URL (which contains <pkgName> and is not a package:// URL itself).
Example:
   base URL: https://example.com/<pkgName>/urdf/robot.urdf
   resolved: https://example.com/<pkgName>/<pkgPath>
```

This allows loading entire URDFs from Github URLs, for instance:

-
https://raw.githubusercontent.com/LeoRover/leo_common-ros2/humble/leo_description/urdf/leo.urdf.xacro
-
https://raw.githubusercontent.com/LeoRover/leo_common-ros2/de35d0ee7/leo_description/urdf/leo.urdf.xacro