Fix a handful of plot issues (#7229)
**User-Facing Changes**
Fix an issue where plotting a signal with the header stamp would result
in misleading output.
Revert a "fix" that was intended to handle missing block data more
safely.
Fix an issue where downsampled plot data points could appear out of
order.

**Description**

This PR fixes ~three~ four separate issues:
1. #7193 caused a regression for topic aliases. I'm just going to revert
it for now.
2. We were incorrectly showing the dots for signals that use header
stamps. I centralized the logic that downsamples entire signals at once
and ensured they correctly detect when we actually can just render the
raw points.
3. (Unrelated to the customer issues) Our downsampler occasionally
reversed the order of points inside of a single interval. We now sort
the points inside of an interval.
4. It turns out that `showLine` can also appear on `TypedDataset`s and
not just on the path, which is news to me. This caused an issue where we
were hiding points (with `pointRadius=0`) that were not connected by a
line.

<!-- link relevant GitHub issues -->
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->