Lower transform-tree history limit from 50k to 10k (#7304)
**User-Facing Changes**
<!-- will be used as a changelog entry -->

**Description**

For websocket connections with many transforms over long periods of time
this limit becomes a bit more untenable, from both memory and speed
perspectives. The larger the coordinate frame history the more time it
takes to insert into the history, thus building up the time it takes to
render a frame, which can back up the websocket. I've lowered this by 5x
to keep this and memory concerns with multiple image panels in check

Expected user behavior:

Transform tree memory usage and performance improved at the cost of a
shorter history for coordinate frame transforms.



<!-- link relevant GitHub issues -->
FG-6239
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->