Fix issues with invalid plot series (#7356)
**User-Facing Changes**
Not worth calling out

**Description**
Fixed bugs with series that are invalid (fail to parse) causing tooltips
and current values to display on a different series. (Similar to #7342
but for invalid series)

Approach: instead of each builder handling its own `data: []` defaults
(added in #7342), the builders use the series' `configIndex` to output
`datasetsByConfigIndex`. The coordinator fills in missing entries with
`data: []` before passing to the renderer.


https://github.com/foxglove/studio/assets/14237/c8f6ba28-8eb9-46d8-a3cc-6b914c7aab93