Fix dragging topic names with special characters to the Image panel (#7359)
**User-Facing Changes**
Fixed an issue with dragging and dropping topics into the Image panel if
the topic names contain special characters.

**Description**
The `path` field is quoted, but the Image panel uses an unquoted topic
name in its config, so drag & drop would result in an invalid config if
the topic name required quotes. This fixes the issue by adding a new
`topicName` field to pass through the unmodified topic name.