Buffer messages from mcap sources as raw Uint8Arrays  (#7376)
**User-Facing Changes**
Store buffered messages as raw typed arrays to reduce memory pressure

**Description**
This PR aims to reduce OOM crashes by changing the playback buffer to
store messages as raw Uint8Arrays. Storing messages as Uint8arrays
reduces heap memory pressure as they do not count towards the heap
memory size compared to how deserialized messages (plain JS objects)
would. A positive consequence of this is, that we can buffer much more
messages than we were previously able to.

Whereas previously, messages would be deserialized in the worker, the
worker now sends (or rather transfers) undeserialized messages to the
main thread where they are stored in the playback buffer and
deserialized on the fly. The only downside of this approach is, that the
main thread now has to pay the cost of message deserialization. However,
this cost might be compensated by more efficient webworker communication
(due to the transferring of array buffers) and is worth to pay in return
for less OOM crashes and a larger playback buffer.

Note that for now the above was only implemented for MCAP related
sources (local/remote mcap files, nuscenes)

Supersedes #7267
Resolves FG-6094