Move wrapPlayer out of useMemo in PlayerManager (#7404)
The `wrapPlayer` call runs in a memo which executes during the _render_
phase in react. However, `wrapPlayer` may run logic which invokes a
state update. React does not like state updates of a component to occur
when another component is rendering. This moves the construction of the
wrapped player to a callback that builds the topicAliasPlayer and the
wrapped player and then sets the state.

This avoids the error for updating a component while rendering.